requirements.txt
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
defusedxml==0.7.1
executing==0.8.3
falcon==3.1.0
gunicorn==20.1.0
httpie==3.2.1
idna==3.3
iniconfig==1.1.1
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
msgpack-python==0.5.6
multidict==6.0.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.29
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
urllib3==1.26.9
waitress==2.1.2
wcwidth==0.2.5
